fix(ios): ensure depositSats is correctly converted to Int64 and improve database operation handling

+++ ios/StableChannels/StableChannels/AppState.swift
@@ -1405,7 +1405,7 @@ class AppState {
 
             let deposit = DepositRecordInput(
                 depositId: depositId,
-                depositSats: depositSats,
+                depositSats: Int64(depositSats),
                 amountUSD: amountUSD,
                 btcPrice: price > 0 ? price : nil
             )
